Fixing append table, normalizing line expansion
Prior to this, dice expanded a bit differently than named lookups. Now
they use the same format (with a minor caveat on table indices).

This involved moving logic into the renderer. It's a bit of a kludge,
but it works.

It also involved debugging the container. The process to register the
table_entry fired multiple times during initialization. Which meant
that the registry I had was different at different points in the
application.

I since switched to ensure that the LineRenderer leverages the same
entry, as does the InputProcessor.